既存のAccessTokenを入力する場合とOAuthで取得する場合とを共存するような処理にした
